requirements.txt
accelerate>=0.21.0
diffusers>=0.19.3
numpy>=1.23.5
pandas>=2.1.1
Pillow>=9.5.0
scikit-learn>=1.0.2
torch>=2.0.0
tqdm>=4.65.0
transformers>=4.30.2